refactor: rename daggers pkg as catalog
