+++ catalog/golang/dagger.go
