+++ catalog/precommit/options.go
