+++ mage/precommit/mage.go
@@ -3,7 +3,7 @@ package precommit
 import (
 	"context"
 
-	precommitdagger "github.com/mesosphere/daggers/dagger/precommit"
+	precommitdagger "github.com/mesosphere/daggers/catalog/precommit"
 )
 
 // Precommit runs all the precommit checks. Run `mage help:precommit` for information on available options.
